pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Iris-Raine <iris.raine.darkest.star@gmail.com>"]
readme = "README.md"
packages = [
  {include = "task_manager"}
]

[tool.poetry.dependencies]
python = "^3.8.1"
django = "^4.1.5"
python-dotenv = "^0.21.1"
dj-database-url = "^1.2.0"
django-bootstrap4 = "^22.3"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.3.0"
django-filter = "^22.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
